+++ project_config/webpack/util/make_plugins.js
@@ -42,6 +42,9 @@ module.exports = (config) => {
     // ProvidePlugin
     if (config.provide) plugins.push(new webpack.ProvidePlugin(config.provide));
 
+    // DefinePlugin
+    if (config.define) plugins.push(new webpack.DefinePlugin(config.define));
+
     plugins.push(new FriendlyErrorsWebpackPlugin());
 
     return plugins;